mk.admin: add man-html target
mandoc -Thtml output is easy to customize, but mandoc itself is not as
available as groff.  Therefore mandoc has to be built first, before all
man pages can be converted to HTML-4.01 strict markup.

The resulting html pages are very similar to the output of man-cgi, an
old chain of sed-scripts that is still widely in use.