Disable module loading in config
Mopherd now has a standard way to disable module loading in mopherd.conf. To do
this you have set the module key in the "modules" hashtable to zero.

mopherd.conf

  modules[p0f] = 0

This line disables the loading of the p0f module. Module keys are derived from
the shared object filename. p0f.so uses the key "p0f", dblog.so "dblog" and so
on.