Added new configuration directive default_database
Mopher has a database configuration for each table. To avoid having the same
table configuration again and again in your mopherd.conf, the new directive
default_database was added.

CAVEAT:
The default_database is referenced when a table has no configuration of its own
an cannot be used to set defaults for particular values such as database user
or port.

mopherd.conf

  default_database = {
    driver      = "postgresql",
    database	= "mopher",
    user	= "mopher",
    pass	= "apple-pie"
  }